+++ src/router/routes.tsx
@@ -150,6 +150,14 @@ const CreateGuestbookPage = lazy(() =>
   )
 )
 
+const EditGuestbookPage = lazy(() =>
+  import('../sections/guestbooks/edit-guestbook/EditGuestbookFactory').then(
+    ({ EditGuestbookFactory }) => ({
+      default: () => EditGuestbookFactory.create()
+    })
+  )
+)
+
 const NotFoundPage = lazy(() =>
   import('../sections/not-found-page/NotFoundPageFactory').then(({ NotFoundPageFactory }) => ({
     default: () => NotFoundPageFactory.create()
@@ -352,6 +360,15 @@ export const routes: RouteObject[] = [
                 element: (
                   <Suspense fallback={<GuestbookSkeleton />}>
                     <CreateGuestbookPage />
+                  </Suspense>
+                ),
+                errorElement: <ErrorPage />
+              },
+              {
+                path: Route.GUESTBOOKS_EDIT,
+                element: (
+                  <Suspense fallback={<GuestbookSkeleton />}>
+                    <EditGuestbookPage />
                   </Suspense>
                 ),
                 errorElement: <ErrorPage />