+++ src/sections/guestbooks/edit-guestbook/EditGuestbook.tsx
@@ -2,6 +2,7 @@ import { useEffect } from 'react'
 import { Trans, useTranslation } from 'react-i18next'
 import { Alert } from '@iqss/dataverse-design-system'
 import { type NavigateFunction, useNavigate } from 'react-router-dom'
+import { toast } from 'react-toastify'
 import { CollectionRepository } from '@/collection/domain/repositories/CollectionRepository'
 import { GuestbookDTO } from '@/guestbooks/domain/useCases/DTOs/GuestbookDTO'
 import { RouteWithParams } from '@/sections/Route.enum'
@@ -9,46 +10,62 @@ import { useCollection } from '@/sections/collection/useCollection'
 import { NotFoundPage } from '@/sections/not-found-page/NotFoundPage'
 import { BreadcrumbsGenerator } from '@/sections/shared/hierarchy/BreadcrumbsGenerator'
 import { useLoading } from '@/shared/contexts/loading/LoadingContext'
+import { useGetGuestbookById } from '@/sections/dataset/dataset-guestbook/useGetGuestbookById'
 import { useGuestbookRepository } from '../GuestbookRepositoryContext'
 import { GuestbookSkeleton } from '../GuestbookSkeleton'
 import { GuestbookForm } from '../guestbook-form/GuestbookForm'
-import { useCreateGuestbook } from './useCreateGuestbook'
+import { useEditGuestbook } from './useEditGuestbook'
 
-interface CreateGuestbookProps {
+interface EditGuestbookProps {
   collectionId: string
+  guestbookId: number
   collectionRepository: CollectionRepository
 }
 
-export const CreateGuestbook = ({ collectionId, collectionRepository }: CreateGuestbookProps) => {
+export const EditGuestbook = ({
+  collectionId,
+  guestbookId,
+  collectionRepository
+}: EditGuestbookProps) => {
   const { t } = useTranslation('guestbooks')
+  const { t: tShared } = useTranslation('shared')
   const navigate: NavigateFunction = useNavigate()
   const guestbookRepository = useGuestbookRepository()
   const { setIsLoading } = useLoading()
   const { collection, isLoading } = useCollection(collectionRepository, collectionId)
+  const { guestbook, isLoadingGuestbook, errorGetGuestbook } = useGetGuestbookById({
+    guestbookRepository,
+    guestbookId
+  })
   const guestbooksGuideUrl =
     'https://guides.dataverse.org/en/latest/user/dataverse-management.html#dataset-guestbooks'
   const navigateToGuestbooks = () => navigate(RouteWithParams.GUESTBOOKS(collectionId))
-  const { isCreatingGuestbook, errorCreatingGuestbook, handleCreateGuestbook } = useCreateGuestbook(
-    {
-      guestbookRepository,
-      collectionIdOrAlias: collectionId,
-      onSuccessfulCreate: navigateToGuestbooks
+  const { isEditingGuestbook, errorEditingGuestbook, handleEditGuestbook } = useEditGuestbook({
+    guestbookRepository,
+    onSuccessfulEdit: () => {
+      toast.success(t('alerts.updated'))
+      navigateToGuestbooks()
     }
-  )
+  })
+  const isLoadingData = isLoading || isLoadingGuestbook
 
   useEffect(() => {
-    setIsLoading(isLoading)
-  }, [isLoading, setIsLoading])
+    setIsLoading(isLoadingData)
+  }, [isLoadingData, setIsLoading])
 
-  const handleSubmit = (guestbook: GuestbookDTO) => {
-    void handleCreateGuestbook(guestbook)
+  const handleSubmit = (guestbookDTO: GuestbookDTO) => {
+    void handleEditGuestbook(guestbookId, guestbookDTO)
   }
 
   if (!isLoading && !collection) {
     return <NotFoundPage dvObjectNotFoundType="collection" />
   }
 
-  if (isLoading || !collection) {
+  if (!isLoadingGuestbook && !guestbook && errorGetGuestbook === null) {
+    return <NotFoundPage />
+  }
+
+  if (isLoadingData || !collection || (!guestbook && errorGetGuestbook === null)) {
     return <GuestbookSkeleton />
   }
 
@@ -57,14 +74,14 @@ export const CreateGuestbook = ({ collectionId, collectionRepository }: CreateGu
       <BreadcrumbsGenerator
         hierarchy={collection.hierarchy}
         withActionItem
-        actionItemText={t('create.title')}
+        actionItemText={t('edit.title')}
         actionItems={[
           {
             text: t('title'),
             url: RouteWithParams.GUESTBOOKS(collectionId)
           },
           {
-            text: t('create.title')
+            text: t('edit.title')
           }
         ]}
       />
@@ -79,15 +96,19 @@ export const CreateGuestbook = ({ collectionId, collectionRepository }: CreateGu
         />
       </Alert>
 
-      {errorCreatingGuestbook && <Alert variant="danger">{errorCreatingGuestbook}</Alert>}
+      {errorGetGuestbook && <Alert variant="danger">{errorGetGuestbook}</Alert>}
+      {errorEditingGuestbook && <Alert variant="danger">{errorEditingGuestbook}</Alert>}
 
-      <GuestbookForm
-        isSubmitting={isCreatingGuestbook}
-        submitButtonText={t('create.submit')}
-        cancelButtonText={t('create.cancel')}
-        onSubmit={handleSubmit}
-        onCancel={navigateToGuestbooks}
-      />
+      {guestbook && (
+        <GuestbookForm
+          initialGuestbook={guestbook}
+          isSubmitting={isEditingGuestbook}
+          submitButtonText={tShared('saveChanges')}
+          cancelButtonText={tShared('cancel')}
+          onSubmit={handleSubmit}
+          onCancel={navigateToGuestbooks}
+        />
+      )}
     </section>
   )
 }